test: Fix failing unit tests - move unit tests using BuildConfig into androidTest
Resolves: DCMAW-7426
