+++ app/src/androidTest/java/uk/gov/onelogin/TestUtils.kt
@@ -20,6 +20,36 @@ object TestUtils {
         )
     )
 
+    val updateRequiredData = AppInfoData(
+        apps = AppInfoData.App(
+            AppInfoData.AppInfo(
+                minimumVersion = "2.0.0",
+                releaseFlags = AppInfoData.ReleaseFlags(
+                    true,
+                    true,
+                    true
+                ),
+                available = true,
+                featureFlags = AppInfoData.FeatureFlags(true)
+            )
+        )
+    )
+
+    val extractVersionErrorData = AppInfoData(
+        apps = AppInfoData.App(
+            AppInfoData.AppInfo(
+                minimumVersion = "vers1.0.0",
+                releaseFlags = AppInfoData.ReleaseFlags(
+                    true,
+                    true,
+                    true
+                ),
+                available = true,
+                featureFlags = AppInfoData.FeatureFlags(true)
+            )
+        )
+    )
+
     fun AndroidComposeTestRule<ActivityScenarioRule<MainActivity>, MainActivity>.setActivity(
         action: () -> Unit
     ) {
